+++ app/classes/helpers/helpers.py
@@ -47,6 +47,7 @@ logger = logging.getLogger(__name__)
 
 PLAIN_TEXT = "text/plain"
 BAT_TEXT = "application/x-bat"
+ERROR_LOG_HIGHLIGHT = r'<span class="mc-log-error">\1</span>'
 
 MASTER_CONFIG = {
     "https_port": 8443,
@@ -146,7 +147,10 @@ class Helpers:
         self.db_path = os.path.join(
             self.root_dir, "app", "config", "db", "crafty.sqlite"
         )
-        self.big_bucket_cache = os.path.join(self.config_dir, "bigbucket.json")
+        self.big_bucket_minecraft_cache = os.path.join(
+            self.config_dir, "bigbucket.json"
+        )
+        self.big_bucket_hytale_cache = os.path.join(self.config_dir, "hytale.json")
         self.credits_cache = os.path.join(self.config_dir, "credits.json")
         self.passhasher = PasswordHasher()
         self.exiting = False
@@ -344,7 +348,7 @@ class Helpers:
             with open(path, "r", encoding="utf-8"):
                 logger.info(f"{path} is readable")
             return True
-        except PermissionError:
+        except (PermissionError, FileNotFoundError):
             return False
 
     @staticmethod
@@ -815,7 +819,8 @@ class Helpers:
         replacements = [
             (r"(\[.+?/INFO\])", r'<span class="mc-log-info">\1</span>'),
             (r"(\[.+?/WARN\])", r'<span class="mc-log-warn">\1</span>'),
-            (r"(\[.+?/ERROR\])", r'<span class="mc-log-error">\1</span>'),
+            (r"(\[.+?/ERROR\])", ERROR_LOG_HIGHLIGHT),
+            (r"(\[.+?/SEVERE\])", ERROR_LOG_HIGHLIGHT),
             (r"(\[.+?/FATAL\])", r'<span class="mc-log-fatal">\1</span>'),
             (
                 r"(\w+?\[/\d+?\.\d+?\.\d+?\.\d+?\:\d+?\])",
@@ -824,7 +829,8 @@ class Helpers:
             (r"\[(\d\d:\d\d:\d\d)\]", r'<span class="mc-log-time">[\1]</span>'),
             (r"(\[.+? INFO\])", r'<span class="mc-log-info">\1</span>'),
             (r"(\[.+? WARN\])", r'<span class="mc-log-warn">\1</span>'),
-            (r"(\[.+? ERROR\])", r'<span class="mc-log-error">\1</span>'),
+            (r"(\[.+? ERROR\])", ERROR_LOG_HIGHLIGHT),
+            (r"(\[.+? SEVERE\])", ERROR_LOG_HIGHLIGHT),
             (r"(\[.+? FATAL\])", r'<span class="mc-log-fatal">\1</span>'),
         ]