+++ app/classes/remote_stats/stats.py
@@ -6,7 +6,7 @@ import datetime
 import base64
 import typing as t
 
-from app.classes.minecraft.mc_ping import ping
+from app.classes.remote_stats.mc_ping import ping
 from app.classes.models.management import HostStats
 from app.classes.models.servers import HelperServers
 from app.classes.shared.null_writer import NullWriter
@@ -124,8 +124,6 @@ class Stats:
                 "mem_total": "",
                 "disk_data": [],
             }
-        # server_stats = self.get_servers_stats()
-        # data['servers'] = server_stats
 
         return {
             "node_stats": node_stats,
@@ -222,7 +220,7 @@ class Stats:
         server_port = server["server_port"]
 
         logger.debug(f"Pinging {internal_ip} on port {server_port}")
-        if HelperServers.get_server_type_by_id(server_id) != "minecraft-bedrock":
+        if HelperServers.get_server_type_by_id(server_id) == "minecraft-java":
             int_mc_ping = ping(internal_ip, int(server_port))
 
             ping_data = {}