+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 ## --- [4.9.0] - 2026/TBD
 ### New features
-TBD
+- Add Support For Hytale Servers ([Merge Request](https://gitlab.com/crafty-controller/crafty-4/-/merge_requests/954))
 ### Bug fixes
 - Add error data to servers running global dir ([Merge Request](https://gitlab.com/crafty-controller/crafty-4/-/merge_requests/961))
 ### Tweaks